tests/locale.sh: remove unportable test with external 'wc' command
The -C flag to 'wc' is not portable; '-m' is the portable variant.
However, this whole test is not appropriate here, as it tests the
external 'wc' utility and not ksh's builtin version -- so any test
failure represents a bug in the OS and not in ksh.

src/cmd/ksh93/tests/locale.sh:
- Remove test that counts two € (euro) character using the OS's
  external 'wc' utility. Keep the test that uses builtin 'wc'.