Avoid brute-forcing during the build process to prevent failed builds
The default behavior of the package script is to ignore
errors. This has disastrous effects if any part of ksh
or libast fails to build due to something like a syntax
error, as the build will stop long after the first error
occurred.

This commit removes the -k flag from the $makeflags variable
in both copies of the package script. This forces the build
to fail if a compiler error occurs.

bin/package and src/cmd/INIT/package.sh:
 - Remove the -k assignment to $makeflags so that compiler
   errors cause the build to fail.